Build Release configuration on Windows (#14)
As documented by @rwhitt2049 in conda-forge/python-xxhash-feedstock#15
the windows build of xxhash is a debug build and links against the
Windows Debug runtime which is not available on most machines.
It can also be seen from the Azure pipelines log[^1] that the recipe
builds the Debug configuration on Windows. According to the CMake
documentation[^2] specifying CMAKE_BUILD_TYPE has no effect on the
Visual Studio generator. Instead a StackOverflow post[^3] says to use
--config Release in the build command.

[^1]: https://dev.azure.com/conda-forge/feedstock-builds/_build/results?buildId=248039&view=logs&j=a70f640f-cc53-5cd3-6cdc-236a1aa90802&t=f5d15007-a01c-5ad8-c9ce-4d519d3b275f&l=403

[^2]: https://cmake.org/cmake/help/latest/variable/CMAKE_BUILD_TYPE.html

[^3]: https://stackoverflow.com/a/20423820